More error checking; esp. exceptions with client
I'm not sure what I was thinking back then. Maybe I had this work on
my mental to do list and then forgot it. But this should detect
exceptions happening on requests, not just error statuses, and abort
when appropriate. Also, made ping more resilient to any errors so
that it'll get to the final report as best it can.