New configuration file option and more.
You can now put all your Swiftly CLI settings into a configuration file
instead of the environment or putting on the command line every time.
Uses ~/.swiftly.conf by default.

In this work I also standardized the treatment of all the options. So
they can all be in the conf file, in the environment, or on the command
line. Each of those overrides the previous.

I also made the Swiftly CLI extendable, since I was working on BrimLock.

Oh, and -? -h --help should work in even more places now.